refactor: improve project directory architecture
