+++ src/erc_tokens_contracts/ERC721PsiLandMinter.sol
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.7;
-import "./utils/AccessProtected.sol";
+import "../utils/AccessProtected.sol";
 import "erc721psi/contracts/ERC721Psi.sol";
 import "@openzeppelin/contracts/utils/Strings.sol";
 import { IERC721Receiver } from "@openzeppelin/contracts/token/ERC721/IERC721Receiver.sol";
