+++ src/staking_contracts/stake.sol
