+++ test/01_ERC721Token.test.ts
@@ -4,7 +4,7 @@ dotenv.config();
 import { setupWallet, zkEVM_provider, ownerSigner, adminSigner } from "./utils/setupWallet";
 import { ethers, Contract } from "ethers";
 import { checkBalances } from "./utils/checkBalances";
-import { abi, bytecode } from "../artifacts/src/ERC721Token.sol/TestTokenERC721.json";
+import { abi, bytecode } from "../artifacts/src/erc_tokens_contracts/ERC721Token.sol/TestTokenERC721.json";
 
 describe("ERC20 tokens deployment & tests on zkEVM", async () => {
     // declare an instance of the contract to be deployed
