+++ test/09_Interface.test.ts
@@ -4,8 +4,8 @@ dotenv.config();
 import { setupWallet, zkEVM_provider, ownerSigner } from "./utils/setupWallet";
 import { ethers, Contract } from "ethers";
 import { checkBalances } from "./utils/checkBalances";
-import counter_artifacts from "../artifacts/src/interface.sol/Counter.json";
-import myContract_artifacts from "../artifacts/src/interface.sol/MyContract.json";
+import counter_artifacts from "../artifacts/src/interface_contracts/interface.sol/Counter.json";
+import myContract_artifacts from "../artifacts/src/interface_contracts/interface.sol/MyContract.json";
 
 describe("Interface contract deployment & tests on zkEVM", async () => {
     // declare an instance of the contract to be deployed
